Remove glue record/NS record reflection
There is actually no case where this is useful.

If the subdomain's NS records are external to the zone as defined in Kubernetes, then the records
are unknowable to the Kubizone controller, unless it does external DNS lookups which is outside
the scope of its functioning.

If the subdomain's NS records are internal, then they are likely the same as the parent domain,
and would therefore not need to be set, or be set manually via custom Records.